backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.36.0
torch==2.5.1
sentencepiece==0.2.0
protobuf==4.25.1
numpy==1.26.4
scipy==1.11.4
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==21.2.0
